Item14412: Don't hardcode the protocol
When a virtual host is set,  the protocol is hardcoded to http://   This
is fine in a normal Foswiki configuration because the "Requests" URI is
used to build links.  However in some reverse proxy configurations, the
client uses https:// to reach the proxy, and http:// is used internally on the last
hop to the server.  In most cases, setting {ForceDefaultUrlHost} will
force the protocol to https.  But not if the VHC has hardcoded it to
http.

Also make sure that configure is usable, even if running on a virtual
host, but only when the virtual host matches the base DefaultUrlHost.